MONDRIAN: Integrate from mondrian-release/3.2 to main @14042. This is the last     tranche of integration from the 3.2 branch. It includes various performance     improvements, and the switch to a JavaCC-based parser. Also fixes a     checkFile exception in Util.
[git-p4: depot-paths = "//open/mondrian/": change = 14043]

+++ src/main/mondrian/olap/Scanner.java
@@ -330,7 +330,7 @@ public class Scanner {
      * @return number literal token
      */
     private Symbol makeNumber(BigDecimal mantissa, int exponent) {
-        double d = mantissa.movePointRight(exponent).doubleValue();
+        BigDecimal d = mantissa.movePointRight(exponent);
         return makeSymbol(ParserSym.NUMBER, d);
     }
